feat: split into app crate
