+++ crates/machete-app/src/apps/library/display.rs
