+++ crates/machete-app/src/settings_panel.rs
