+++ crates/machete-app/src/ui_models/filters.rs
@@ -3,8 +3,7 @@ use std::{
     collections::HashMap,
     hash::{Hash, Hasher},
 };
-
-use crate::models::{
+use machete::models::{
     ids::InternalId,
     library::{
         creature::{Alignment, Size},
