+++ crates/machete-app/src/ui_models/mod.rs
