+++ crates/machete-app/src/update_context.rs
