+++ crates/machete-app/src/widgets/hidden_combo_box.rs
