+++ crates/machete-app/src/widgets/restricted_text_edit.rs
